ticketsystem/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
gunicorn==23.0.0
fpdf2==2.8.5
Flask-WTF==1.2.2
qrcode[pil]==8.2
Flask-Limiter==4.1.1
flask-talisman==1.1.0
Flask-APScheduler==1.13.1
pyOpenSSL==24.0.0
prometheus-client==0.20.0
